Remove Exec. Bit
Removes executable attributes on files that clearly do not need such.

The purpose of of this is to make certain files that are not intended to be executable
cannot be run, no matter what their type.

It also makes checking for malware/viruses and/or other types of unwanted files alot easier/faster

png files do not deserve the executable bit, so same for xml.